feat(keymaps): implemented more detailed descriptions
By default all keymaps got the description `blink.cmp` for someone new
to the keymaps (like me) it's very helpful to with a tool like
Telescope be able to quickly look up what keymaps do. Hence the
description is very important to provide the detail.

Logic has been added to generate a description in the format:
```
Blink: command_1, command_2, ... (mode)
```

This is just a very quick solution implemented which is good enough for
me, as far as im aware this change is self contained and affects
essentially nothing other than the descriptions.